+++ apps/desktop/src/renderer/screens/main/components/WorkspaceSidebar/PortsList/components/WorkspacePortGroup/index.ts
@@ -1,0 +1,1 @@
+export { WorkspacePortGroup } from "./WorkspacePortGroup";
